+++ nflow-jetty/src/test/java/io/nflow/jetty/mapper/StateVariableValueTooLongExceptionMapperTest.java
@@ -1,4 +1,4 @@
-package io.nflow.rest.mapper;
+package io.nflow.jetty.mapper;
 
 import static javax.ws.rs.core.Response.Status.BAD_REQUEST;
 import static org.hamcrest.CoreMatchers.is;
@@ -8,14 +8,16 @@ import javax.ws.rs.core.Response;
 
 import org.junit.jupiter.api.Test;
 
+import io.nflow.engine.workflow.executor.StateVariableValueTooLongException;
+import io.nflow.jetty.mapper.StateVariableValueTooLongExceptionMapper;
 import io.nflow.rest.v1.msg.ErrorResponse;
 
-public class IllegalArgumentExceptionMapperTest {
-  IllegalArgumentExceptionMapper mapper = new IllegalArgumentExceptionMapper();
+public class StateVariableValueTooLongExceptionMapperTest {
+  StateVariableValueTooLongExceptionMapper mapper = new StateVariableValueTooLongExceptionMapper();
 
   @Test
   public void exceptionIsMappedToBadRequest() {
-    try (Response response = mapper.toResponse(new IllegalArgumentException("error"))) {
+    try (Response response = mapper.toResponse(new StateVariableValueTooLongException("error"))) {
       assertThat(response.getStatus(), is(BAD_REQUEST.getStatusCode()));
       ErrorResponse error = (ErrorResponse) response.getEntity();
       assertThat(error.error, is("error"));